Revert "Revert "Merge branch 'main' into Adam's-Branch""
This reverts commit b669027f5b796ec7847aa38379c5004bdbb95910.

+++ GroceryStore-Backend/src/main/java/ca/mcgill/ecse321/grocerystore/dao/DeliveryPersonRepository.java
@@ -7,6 +7,6 @@ import ca.mcgill.ecse321.grocerystore.model.DeliveryPerson;
 
 public interface DeliveryPersonRepository extends CrudRepository<DeliveryPerson, Integer> {
 	
-	DeliveryPerson findByRoleID(Integer roleID);
+	DeliveryPerson findByRoleID(Long roleID);
 	
 }
