+++ GroceryStore-Backend/src/main/java/ca/mcgill/ecse321/grocerystore/dao/StoreRepository.java
@@ -5,6 +5,5 @@ import ca.mcgill.ecse321.grocerystore.model.Store;
 
 public interface StoreRepository extends CrudRepository<Store,Integer>{
 
-		Store findByStoreID(Integer storeID);
-		
+		Store findByStoreID(Long storeID);	
 }
